# Fix stopping criterion of _graph_connected_components