--- a/sklearn/manifold/spectral_embedding_.py
+++ b/sklearn/manifold/spectral_embedding_.py
@@ -42,24 +42,23 @@ def _graph_connected_component(graph, node_id):
         belonging to the largest connected components of the given query
         node
     """
-    connected_components_matrix = np.zeros(
-        shape=(graph.shape[0]), dtype=np.bool)
-    nodes_to_explore = np.zeros(shape=(graph.shape[0]), dtype=np.bool)
-    nodes_to_explore[node_id] = True
     n_node = graph.shape[0]
-    for i in range(n_node):
-        last_num_component = connected_components_matrix.sum()
-        np.logical_or(connected_components_matrix,
+    connected_nodes = np.zeros(
+        shape=(n_node), dtype=np.bool)
+    nodes_to_explore = np.zeros(shape=(n_node), dtype=np.bool)
+    nodes_to_explore[node_id] = True
+    for _ in range(n_node):
+        last_num_component = connected_nodes.sum()
+        np.logical_or(connected_nodes,
                       nodes_to_explore,
-                      connected_components_matrix)
+                      out=connected_nodes)
         indices = np.where(nodes_to_explore)[0]
         nodes_to_explore.fill(False)
         for i in indices:
-            np.logical_or(nodes_to_explore, graph[i], nodes_to_explore)
-
-        if last_num_component >= connected_components_matrix.sum():
+            np.logical_or(nodes_to_explore, graph[i], out=nodes_to_explore)
+        if last_num_component >= connected_nodes.sum():
             break
-    return connected_components_matrix
+    return connected_nodes
 
 
 def _graph_is_connected(graph):