The function didn't stop in the case of a cyclic graph. I restored the previous stopping criterion and left the optimization that only bring a small overhead on my box (50ms on 10 tries).

Related to #5639