@giorgiop 2015-11-05
Could you check the difference with master in runtime when you test `manifold/tests/test_spectral_embedding` and the whole code base?