@AlexandreAbraham 2015-11-06
Before any optimization:
```
$ nosetests sklearn/manifold/tests/test_spectral_embedding.py --pdb
/home/aa013911/mywork/scikit-learn/sklearn/__check_build/__init__.py:44: RuntimeWarning: compiletime version 2.6 of module 'sklearn.check_build._check_build' does not match runtime version 2.7
  from ._check_build import check_build
...S.....
----------------------------------------------------------------------
Ran 9 tests in 2.173s

OK (SKIP=1)
```
After first PR
```
$ nosetests sklearn/manifold/tests/test_spectral_embedding.py
/home/aa013911/mywork/scikit-learn/sklearn/__check_build/__init__.py:44: RuntimeWarning: compiletime version 2.6 of module 'sklearn.check_build._check_build' does not match runtime version 2.7
  from ._check_build import check_build
...S.....
----------------------------------------------------------------------
Ran 9 tests in 27.387s

OK (SKIP=1)
```
After fixing the stopping criterion
```
$ nosetests sklearn/manifold/tests/test_spectral_embedding.py
/home/aa013911/mywork/scikit-learn/sklearn/__check_build/__init__.py:44: RuntimeWarning: compiletime version 2.6 of module 'sklearn.check_build._check_build' does not match runtime version 2.7
  from ._check_build import check_build
...S......
----------------------------------------------------------------------
Ran 10 tests in 2.201s

OK (SKIP=1)
```

It is slightly slower but it uses a fix amount of memory as opposite to the old version. I have tried to use indices instead of boolean vectors and it is slightly slower (about 70ms) but I can use this option if needed.