@AlexandreAbraham 2015-11-08
@giorgiop this one should be good to merge.