@giorgiop 2015-11-08
Did you try to see if we can completely avoid the allocation of the array `nodes_to_explore`?