@giorgiop 2015-11-13
I am running this script to measure gain in runtime:
```
import numpy as np
from time import time
from sklearn.manifold.spectral_embedding_ import _graph_connected_component


def perturb_graph(graph):
    m = graph.shape[0]
    graph[rng.randint(m, size=2 * m),
          rng.randint(m, size=2 * m)] = 1

n = 1000
rng = np.random.RandomState(12)

graph = np.zeros(shape=(n, n))

start = time()
for _ in range(20):
    perturb_graph(graph)  # perturb graph at random
    init = rng.randint(n)
    graph[0, init] = 1  # starting node
    _graph_connected_component(graph, init)
end = time() - start
print("time: %.3f" % end)
```

```
# on master
time: 59.612
# this branch
time: 0.303
# above + moving up the break condition
time: 0.237
# above + no fill `nodes_to_explore` with False
time: 0.027
```

Here the final code. I am still not sure if we can completely avoid `nodes_to_explore` but the speed up should be fine already.
```
def _graph_connected_component(graph, node_id):
    """Find the largest graph connected components that contains one
    given node

    Parameters
    ----------
    graph : array-like, shape: (n_samples, n_samples)
        adjacency matrix of the graph, non-zero weight means an edge
        between the nodes

    node_id : int
        The index of the query node of the graph

    Returns
    -------
    connected_components_matrix : array-like, shape: (n_samples,)
        An array of bool value indicating the indexes of the nodes
        belonging to the largest connected components of the given query
        node
    """
    n_node = graph.shape[0]
    connected_nodes = np.zeros(
        shape=(n_node), dtype=np.bool)
    nodes_to_explore = np.zeros(shape=(n_node), dtype=np.bool)
    nodes_to_explore[node_id] = True
    for _ in range(n_node):
        last_num_component = connected_nodes.sum()
        np.logical_or(connected_nodes,
                      nodes_to_explore,
                      out=connected_nodes)
        if last_num_component >= connected_nodes.sum():
            break
        indices = np.where(nodes_to_explore)[0]
        for i in indices:
            np.logical_or(np.zeros(shape=(n_node), dtype=np.bool), graph[i],
                          out=nodes_to_explore)
    return connected_nodes
```