@amueller 2016-10-08
closing as #6268 was merged.